requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
cleantext==1.1.4
click==8.1.7
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.4
executing==2.0.1
filelock==3.13.1
fonttools==4.45.0
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gdown==4.7.1
hjson==3.1.0
huggingface-hub==0.19.4
hyperopt==0.2.7
idna==3.4
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lightning==2.1.2
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.1.1
torchmetrics==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.28.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
xgboost==2.0.2
yarl==1.9.3
